Add files via upload
This HTML document creates a simple registration form. It includes fields for entering a name and password, selecting a class (11 or 12), choosing favorite subjects (Math, Physics, Chemistry, Biology), selecting a city from a dropdown list (Lahore, Karachi, Multan), and providing feedback. The form includes a YouTube video and an embedded Wikipedia page for additional content. The page ends with a footer displaying the text "---page ended---"